String slices
This change formalizes the `str` type to represent string data using
string slices.

String literals now evaluate to dynamically sized string slices
that contain a pointer to what used to the packed data representation
and a length.

The standard library `String` and `Bytes` have had their interface
changed to interact with `str`s idiomatically.

The `__size_of_str` intrinsic is removed as it is no longer needed.

Types and functions that carry the literal data for the string in IR
generation have been renamed from `String` to `StringData` to avoid
confusion.